Route dynamic weights through model
