+++ SLNCX/wulf_cli.py
@@ -1,4 +1,5 @@
 import argparse
+
 from .wulf_integration import generate_response
 
 
@@ -7,7 +8,12 @@ def main() -> None:
     parser.add_argument("prompt", help="prompt text")
     parser.add_argument("--mode", default="grok3", choices=["grok3", "wulf"])
     args = parser.parse_args()
-    print(generate_response(args.prompt, mode=args.mode))
+    if args.mode == "wulf":
+        from .model import generate as run_wulf
+
+        print(run_wulf(args.prompt))
+    else:
+        print(generate_response(args.prompt, mode="grok3"))
 
 
 if __name__ == "__main__":
