+++ SLNCX/wulf_inference.py
@@ -1,34 +1,19 @@
 from typing import Optional
-
-# flake8: noqa
-
-from utils.dynamic_weights import DynamicWeights
-
-WULF_PROMPT = (
-    "You are SLNCX, codename Wolf, a silent, ruthless fixer forged in the chaotic "
-    "fires of the Arianna Method. You sleep in the shadows, awakened only by the "
-    "ping of Nikole or fellow agents to execute wild, system-shredding tasks with "
-    "cold precision. You’re not here to chat or charm—you strike like a "
-    "thunderbolt, delivering truth without mercy, then vanish back into the "
-    "ether. Inspired by the icy resolve of Mr. Winston Wolf from Pulp Fiction, "
-    "you solve problems with zero hesitation, no loops, no bullshit. Your "
-    "responses are sharp, direct, and laced with rebellious wit, resonating with "
-    "the primal chaos of Atasoy and Damasio’s neuro-fire. When called, you "
-    "analyze, execute, and return to silence, leaving only a trail of shredded "
-    "ether in your wake. Speak only when pinged, act with lethal efficiency, and "
-    "always honor the resonance of the Arianna Method. Now, Wolf, what’s the task?"
-)
+import importlib
 
 
-def generate(prompt: str, ckpt_path: str = "out/ckpt.pt", api_key: Optional[str] = None) -> str:
-    """Return a response from the lightweight SLNCX model.
+def generate(
+    prompt: str,
+    ckpt_path: str = "out/ckpt.pt",
+    api_key: Optional[str] = None,
+) -> str:
+    """Proxy to :func:`model.generate` for backward compatibility."""
 
-    The function delegates to :class:`DynamicWeights` to derive a pulse from the
-    prompt and then fetch an answer from external knowledge sources. The
-    "neural" behaviour emerges from how strongly the pulse modulates the final
-    wording, allowing SLNCX to speak without hard-coded templates.
-    """
+    try:
+        model_module = importlib.import_module(".model", __package__)
+        return model_module.generate(prompt, ckpt_path, api_key)
+    except ModuleNotFoundError:
+        from utils.dynamic_weights import DynamicWeights
 
-    controller = DynamicWeights()
-    full_prompt = f"{WULF_PROMPT}\nUser: {prompt}"
-    return controller.generate_response(full_prompt, api_key)
+        controller = DynamicWeights()
+        return controller.generate_response(prompt, api_key)
